change part of paramters types commit for backup only

+++ encode_new.h
@@ -5,8 +5,8 @@ void scalefactor_calc_new (INT32 sb_sample[][3][SCALE_BLOCK][SBLIMIT],
 
 INLINE double mod (double a);
 
-void combine_LR_new (double sb_sample[2][3][SCALE_BLOCK][SBLIMIT],
-		     double joint_sample[3][SCALE_BLOCK][SBLIMIT], int sblimit);
+void combine_LR_new (INT32 sb_sample[2][3][SCALE_BLOCK][SBLIMIT],
+		     INT32 joint_sample[3][SCALE_BLOCK][SBLIMIT], int sblimit);
 
 void find_sf_max (UINT32 sf_index[2][3][SBLIMIT], frame_info * frame,
 		  INT32 sf_max[2][SBLIMIT]);
