+++ psycho_0.h
@@ -1,1 +1,1 @@
-void psycho_0(double SMR[2][SBLIMIT], int nch, unsigned int scalar[2][3][SBLIMIT], FLOAT sfreq);
+void psycho_0(INT32 SMR[2][SBLIMIT], int nch, unsigned int scalar[2][3][SBLIMIT], INT32 sfreq);
